Simplify the test harness considerably
The Reference tree is now just built by walking the Protos
directory and compiling everything there.  We no longer
have or need a list of test protos in the Makefile.

The _test tree is built exactly the same way, followed
by a `diff -r` to compare the _test tree against the
Reference tree.

In particular, it's now trivial to add new test protos
whenever we find issues.